refactor(dom/cascade): split 1,973-line file into 4 focused modules
Split monolithic cascade.rs into:
- specificity.rs (106 lines): CSS specificity calculation
- resolver.rs (521 lines): Selector matching and style resolution
- merge.rs (68 lines): Style merging trait
- tests.rs (1,260 lines): All 70 cascade tests
- mod.rs (19 lines): Public exports

File reduced from 1,973 to 1,974 lines (modularized into 5 files).

Benefits:
- Single Responsibility: Each module has one clear purpose
- Easier navigation: Smaller, focused files
- Better testability: Tests separated into dedicated module
- Maintains backward compatibility: All public APIs unchanged
